09/11/23

+++ src/Ereditarieta/OP/OperaDarte.java
@@ -12,8 +12,7 @@ public abstract class OperaDarte {
         artista = art;
     }
 
-    public abstract float printIngombro();
-
+    public abstract double printIngombro();
 
     public String getTitolo() {
         return titolo;
